LayoutTests/fast/events/key-events-in-input-button-expected.txt

To test manually, press keys and compare results to other browsers.


target - type - ctrlKey,altKey,shiftKey,metaKey - key - code - keyIdentifier - keyCode - charCode - which
Space:
INPUT - keydown - false,false,false,false - - Space - U+0020 - 32 - 0 - 32
INPUT - keypress - false,false,false,false - - Space - - 32 - 32 - 32
INPUT - keyup - false,false,false,false - - Space - U+0020 - 32 - 0 - 32
INPUT - click
Enter:
INPUT - keydown - false,false,false,false - Enter - Enter - Enter - 13 - 0 - 13
INPUT - keypress - false,false,false,false - Enter - Enter - - 13 - 13 - 13
INPUT - click
INPUT - keyup - false,false,false,false - Enter - Enter - Enter - 13 - 0 - 13
A:
INPUT - keydown - false,false,true,false - A - KeyA - U+0041 - 65 - 0 - 65
INPUT - keypress - false,false,true,false - A - KeyA - - 65 - 65 - 65
INPUT - keyup - false,false,true,false - A - KeyA - U+0041 - 65 - 0 - 65
